pyproject.toml
[project]
name = "domain-name-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "exrex>=0.12.0",
    "python-whois>=0.9.6",
]
